requirements.txt
colorlog==3.1.4
matplotlib>=1.5.1
numpy>=1.14.2
python-dateutil==2.5.1
requests==2.9.1
scipy>=0.17.0
